Added templates, changed data access, and changed documentation
