+++ Simulation/6DOF_RK4/properties/configs/legacy_properties.yaml
@@ -1,0 +1,33 @@
+<<: *pysim_simulation
+desired_apogee: 4572
+rocket:
+  <<: *simulated_rocket
+  rasaero_lookup_file: "../lookup/RASAero.csv"
+  C_d: 0.5
+  C_d_s: 1.2
+  length: 3.34
+  dry_mass: 14.691
+  radius: 0.0508
+  cm: !numpy/array [1.48, 0., 0.]
+motor:
+  <<: *motor_M2500
+simulated_body:
+  <<: *simulated_body
+  cm: !numpy/array [1.03, 0., 0.]
+  cp: !numpy/array [0.63, 0., 0.]
+flaps:
+  <<: *simulated_flaps
+  max_ext_length: .0178
+  max_ext_spd: 0.001
+sensors:
+  high_g:
+    RMS: 1.9 # milli g
+  gyro:
+    RMS: 75 # mdps (millidegrees per second)
+  barometer:
+    RMS: .012  # Pascal Conversion
+  bno:
+    error: !eval 2.5/3
+meta:
+  output_file: '../output/simulated_6dof.csv'
+  monte_carlo_output_folder: '../output/monte_carlo'
